added new dependency: SweetAlert2; changed DELETE to add confirmation popup

+++ server/public/scripts/client.js
@@ -1,40 +1,39 @@
-console.log( 'js' );
+console.log('js');
 
-$( document ).ready( function(){
-  console.log( 'JQ' );
-  // Establish Click Listeners
-  setupClickListeners()
-  // load existing koalas on page load
-  getKoalas();
-
+$(document).ready(function() {
+	console.log('JQ');
+	// Establish Click Listeners
+	setupClickListeners();
+	// load existing koalas on page load
+	getKoalas();
 }); // end doc ready
 
 function setupClickListeners() {
-  $( '#addButton' ).on( 'click', function(){
-    console.log( 'in addButton on click' );
-    // creates object with input values from DOM
-    let koalaToSend = {
-      name: $('#nameIn').val(),
-      age: $('#ageIn').val(),
-      gender: $('#genderIn').val(),
-      readyForTransfer: $('#readyForTransferIn').val(),
-      notes: $('#notesIn').val(),
-    };
-    // call saveKoala with the new obejct
-    saveKoala( koalaToSend );
-  }); 
-  $(document).on('click', '.deleteButton', deleteKoala)
+	$('#addButton').on('click', function() {
+		console.log('in addButton on click');
+		// creates object with input values from DOM
+		let koalaToSend = {
+			name: $('#nameIn').val(),
+			age: $('#ageIn').val(),
+			gender: $('#genderIn').val(),
+			ready_to_transfer: $('#readyForTransferIn').val(),
+			notes: $('#notesIn').val()
+		};
+		// call saveKoala with the new obejct
+		saveKoala(koalaToSend);
+	});
+	$(document).on('click', '.deleteButton', deleteKoala);
 }
 
-function getKoalas(){
-  console.log( 'in getKoalas' );
-  // ajax call to server to get koalas
-  $.ajax({
+function getKoalas() {
+	console.log('in getKoalas');
+	// ajax call to server to get koalas
+	$.ajax({
 		type: 'GET',
 		url: '/koalas'
 	})
 		.then(function(response) {
-      console.log('getting koalas:', response)
+			console.log('getting koalas:', response);
 			renderKoalas(response);
 		})
 		.catch(function(error) {
@@ -42,10 +41,10 @@ function getKoalas(){
 		});
 } // end getKoalas
 
-function saveKoala( newKoala ){
-  console.log( 'in saveKoala', newKoala );
-  // ajax to send new koala
-  $.ajax({
+function saveKoala(newKoala) {
+	console.log('in saveKoala', newKoala);
+	// ajax to send new koala
+	$.ajax({
 		type: 'POST',
 		url: '/koalas',
 		data: newKoala
@@ -60,23 +59,38 @@ function saveKoala( newKoala ){
 		});
 }
 
-function deleteKoala(){
-  $.ajax({
-    type: 'DELETE',
-    url: `/koalas/${$(this).parents('tr').data('id')}`
-}).then((res) => {
-    console.log('DELETE:', res);
-    getKoalas();
-}).catch((err) => {
-    console.log('FAILED:', err);
-});
+function deleteKoala() {
+	Swal.fire({
+		title: 'Are you sure?',
+		text: "You won't be able to revert this!",
+		icon: 'warning',
+		showCancelButton: true,
+		confirmButtonColor: '#3085d6',
+		cancelButtonColor: '#d33',
+		confirmButtonText: 'Yes, delete it!'
+	}).then((result) => {
+		if (result.isConfirmed) {
+			$.ajax({
+				type: 'DELETE',
+				url: `/koalas/${$(this).parents('tr').data('id')}`
+			})
+				.then((res) => {
+					console.log('DELETE:', res);
+					getKoalas();
+				})
+				.catch((err) => {
+					console.log('FAILED:', err);
+				});
+			Swal.fire('Deleted!', 'Your file has been deleted.', 'success');
+		}
+	});
 }
 
-function renderKoalas(koalas){
-  console.log(koalas);
-  $('#viewKoalas').empty()
-  for(let koala of koalas){
-    $('#viewKoalas').append(`
+function renderKoalas(koalas) {
+	console.log(koalas);
+	$('#viewKoalas').empty();
+	for (let koala of koalas) {
+		$('#viewKoalas').append(`
     <tr data-id = "${koala.id}">
       <td>${koala.name}</td>
       <td>${koala.age}</td>
@@ -88,5 +102,5 @@ function renderKoalas(koalas){
       </td>
     </tr>
     `);
-  }
+	}
 }
